added gravity
With that there was a little problem with synchronuous moving of multiple objects.
First draft was with Controls on object (quite natural for me to object being responsible for controling it's position).
This again made the objects to "jump" with desynchronisation of movement/drawing cycles between objects.
So I switched to single AppState responsible for controlling positions of every object in system.

Now appreciate the beauty of two objects positioned in cosmic scale (distance > 100M from 0,0,0) each separately influenced by gravity, yet moving in perfect unison :)