work-in-progress of auto-flight
I wanted to add an action like 'fly me to the moon'. This turned out to be little challenge - I have to make some deeper analysis (having object A flying in one direction and object B flying in different direction, which way should object A accelerate to land in distance D from object B with relative velocity of zero?), algorithm so far is almost ready, have to tinker with it a little (last time it crashed right into the Moon - on the upside, it did try to decelerate at the moment).